Update range unit tests
