+++ src/Qir/Runtime/stdlib/src/arrays.rs
@@ -141,7 +141,6 @@ mod tests {
     use std::{mem::size_of, convert::TryInto};
 
     use super::*;
-    use crate::range_support::{quantum__rt__array_slice_1d, Range};
 
     #[test]
     fn test_array_creation_simple() {
@@ -261,94 +260,6 @@ mod tests {
             assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 4), 31);
             assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 5), 0);
             __quantum__rt__array_update_reference_count(arr3, -1);
-        }
-    }
-
-    #[test]
-    fn test_array_slicing() {
-        let arr = __quantum__rt__array_create_1d(1, 3);
-        unsafe {
-            assert_eq!(__quantum__rt__array_get_size_1d(arr), 3);
-            let first = __quantum__rt__array_get_element_ptr_1d(arr, 0);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr, 0), 0);
-            *first = 42;
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr, 0), 42);
-            let second = __quantum__rt__array_get_element_ptr_1d(arr, 1);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr, 1), 0);
-            *second = 31;
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr, 1), 31);
-            let arr2 = __quantum__rt__array_copy(arr, true);
-            assert_eq!(__quantum__rt__array_get_size_1d(arr2), 3);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr2, 0), 42);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr2, 1), 31);
-            let arr3 = __quantum__rt__array_concatenate(arr, arr2);
-            assert_eq!(__quantum__rt__array_get_size_1d(arr3), 6);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 0), 42);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 1), 31);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 2), 0);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 3), 42);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 4), 31);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 5), 0);
-            // Third array crated via concatenation has contents [42, 31, 0, 42, 31, 0], create
-            // fourth array via slicing with step size 2, expected contents [42, 0, 31].
-            let arr4 = quantum__rt__array_slice_1d(
-                arr3,
-                Range {
-                    start: 0,
-                    step: 2,
-                    end: 5,
-                },
-            );
-            assert_eq!(__quantum__rt__array_get_size_1d(arr4), 3);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr4, 0), 42);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr4, 1), 0);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr4, 2), 31);
-            // Create fifth array via slicing with reverse iteration, expected contents
-            // [31, 0, 42].
-            let arr5 = quantum__rt__array_slice_1d(
-                arr3,
-                Range {
-                    start: 4,
-                    step: -2,
-                    end: 0,
-                },
-            );
-            assert_eq!(__quantum__rt__array_get_size_1d(arr5), 3);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr5, 0), 31);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr5, 1), 0);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr5, 2), 42);
-            // Create sixth array with range end less than range start, should succeed and create
-            // an empty array.
-            let arr6 = quantum__rt__array_slice_1d(
-                arr5,
-                Range {
-                    start: 0,
-                    step: 1,
-                    end: -1,
-                },
-            );
-            // Confirm each copy, concatenation, and slice is independent of others.
-            assert_eq!(__quantum__rt__array_get_size_1d(arr6), 0);
-            __quantum__rt__array_update_reference_count(arr, -1);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr2, 1), 31);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 0), 42);
-            __quantum__rt__array_update_reference_count(arr2, -1);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 0), 42);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 1), 31);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 2), 0);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 3), 42);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 4), 31);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr3, 5), 0);
-            __quantum__rt__array_update_reference_count(arr3, -1);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr4, 0), 42);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr4, 1), 0);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr4, 2), 31);
-            __quantum__rt__array_update_reference_count(arr4, -1);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr5, 0), 31);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr5, 1), 0);
-            assert_eq!(*__quantum__rt__array_get_element_ptr_1d(arr5, 2), 42);
-            __quantum__rt__array_update_reference_count(arr5, -1);
-            __quantum__rt__array_update_reference_count(arr6, -1);
         }
     }
